Xong Buoi 2

+++ WebBan_1398/Repositories/EFCategoryRepository.cs
@@ -31,8 +31,11 @@ using WebBan_1398.Models;
         public async Task DeleteAsync(int id)
         {
             var category = await _context.Categories.FindAsync(id);
-            _context.Categories.Remove(category);
-            await _context.SaveChangesAsync();
+            if (category != null)
+            {
+                _context.Categories.Remove(category);
+                await _context.SaveChangesAsync();
+            }
         }
     }
 }
